+++ one_for_all_generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next release
 - feat: support sealed classes
+- refactor!: renamed PlatformException to PlatformError
+- refactor!: the "Result" objects accept on `error` method a `PlatformError` class
 
 ## 1.0.1
 - fix: fixes `DateTime` swift serialization on swift code